fix: address review comments on miner notify auth
- reject empty or whitespace-padded auth tokens at startup
- log and skip notify instead of panicking on invalid header value
- add WWW-Authenticate to 401 and Allow to 405 responses
- collect notify body via req.into_body()
- compare bearer tokens in constant time
- move notify unit tests into standalone tests module
