+++ Cargo.toml
@@ -296,6 +296,7 @@ sled = "0.34.7"
 snap = "1"
 sql-builder = "3.1"
 sqlx = "0.8.2"
+subtle = "2.6"
 syn = "2.0"
 tera = "1"
 thiserror = "1.0.22"
