+++ miner/src/client/mod.rs
@@ -33,10 +33,14 @@ use std::net::SocketAddr;
 use std::sync::Arc;
 use std::sync::atomic::{AtomicU64, Ordering};
 use std::{convert::Into, time};
+use subtle::ConstantTimeEq;
 use tokio::{
     net::TcpListener,
     sync::{mpsc, oneshot},
 };
+
+#[cfg(test)]
+mod tests;
 
 type RpcRequest = (oneshot::Sender<Result<Bytes, RpcError>>, MethodCall);
 
@@ -374,6 +378,7 @@ where
     if req.method() != hyper::Method::POST {
         return Ok(Response::builder()
             .status(hyper::StatusCode::METHOD_NOT_ALLOWED)
+            .header(hyper::header::ALLOW, "POST")
             .body(Empty::new())?);
     }
 
@@ -383,17 +388,18 @@ where
             .get(hyper::header::AUTHORIZATION)
             .and_then(|value| value.to_str().ok())
             .and_then(|value| value.strip_prefix("Bearer "))
-            .map(|token| token.trim())
-            == Some(expected_token.as_str());
+            .map(str::trim)
+            .is_some_and(|token| bool::from(token.as_bytes().ct_eq(expected_token.as_bytes())));
 
         if !authorized {
             return Ok(Response::builder()
                 .status(hyper::StatusCode::UNAUTHORIZED)
+                .header(hyper::header::WWW_AUTHENTICATE, "Bearer")
                 .body(Empty::new())?);
         }
     }
 
-    let body = BodyExt::collect(req).await?.aggregate();
+    let body = BodyExt::collect(req.into_body()).await?.aggregate();
 
     if let Ok(template) = serde_json::from_reader(body.reader()) {
         client.update_block_template(template);
@@ -424,103 +430,5 @@ fn parse_authorization(url: &Uri) -> Option<HeaderValue> {
         Some(header)
     } else {
         None
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use ckb_async_runtime::new_background_runtime;
-    use ckb_channel::bounded;
-    use ckb_jsonrpc_types::BlockTemplate;
-    use http_body_util::Full;
-    use hyper::body::Bytes;
-
-    fn test_client(auth_token: Option<String>) -> (Client, ckb_channel::Receiver<Works>) {
-        let runtime = new_background_runtime();
-        let (new_work_tx, new_work_rx) = bounded::<Works>(16);
-        let config = MinerClientConfig {
-            rpc_url: "http://127.0.0.1:8114/".to_string(),
-            poll_interval: 1000,
-            block_on_submit: false,
-            listen: None,
-            auth_token,
-        };
-        let client = Client::new(new_work_tx, config, runtime);
-        (client, new_work_rx)
-    }
-
-    fn notify_request(
-        method: &str,
-        auth_header: Option<&str>,
-        body: Bytes,
-    ) -> Request<Full<Bytes>> {
-        let mut builder = Request::builder()
-            .method(method)
-            .uri("http://127.0.0.1:8888/");
-        if let Some(header) = auth_header {
-            builder = builder.header(hyper::header::AUTHORIZATION, header);
-        }
-        builder.body(Full::new(body)).unwrap()
-    }
-
-    #[tokio::test]
-    async fn notify_rejects_non_post_requests() {
-        let (client, _rx) = test_client(None);
-        let req = notify_request("GET", None, Bytes::new());
-        let resp = handle(client, req).await.unwrap();
-        assert_eq!(resp.status(), hyper::StatusCode::METHOD_NOT_ALLOWED);
-    }
-
-    #[tokio::test]
-    async fn notify_accepts_unauthenticated_requests_when_no_token_configured() {
-        let (client, _rx) = test_client(None);
-        let req = notify_request("POST", None, Bytes::new());
-        let resp = handle(client, req).await.unwrap();
-        assert_eq!(resp.status(), hyper::StatusCode::OK);
-    }
-
-    #[tokio::test]
-    async fn notify_rejects_missing_token_when_auth_configured() {
-        let (client, _rx) = test_client(Some("secret".to_string()));
-        let req = notify_request("POST", None, Bytes::new());
-        let resp = handle(client, req).await.unwrap();
-        assert_eq!(resp.status(), hyper::StatusCode::UNAUTHORIZED);
-    }
-
-    #[tokio::test]
-    async fn notify_rejects_wrong_token_when_auth_configured() {
-        let (client, _rx) = test_client(Some("secret".to_string()));
-        let req = notify_request("POST", Some("Bearer wrong"), Bytes::new());
-        let resp = handle(client, req).await.unwrap();
-        assert_eq!(resp.status(), hyper::StatusCode::UNAUTHORIZED);
-    }
-
-    #[tokio::test]
-    async fn notify_accepts_correct_token_and_updates_work() {
-        let (client, rx) = test_client(Some("secret".to_string()));
-        let template = BlockTemplate {
-            work_id: 42.into(),
-            ..Default::default()
-        };
-        let body = serde_json::to_vec(&template).unwrap();
-        let req = notify_request("POST", Some("Bearer secret"), Bytes::from(body));
-        let resp = handle(client.clone(), req).await.unwrap();
-        assert_eq!(resp.status(), hyper::StatusCode::OK);
-        assert_eq!(client.current_work_id.load(Ordering::SeqCst), 42);
-
-        let work = rx.recv();
-        assert!(
-            matches!(work, Ok(Works::New(_))),
-            "expected new work notification"
-        );
-    }
-
-    #[tokio::test]
-    async fn notify_accepts_correct_token_with_leading_whitespace() {
-        let (client, _rx) = test_client(Some("secret".to_string()));
-        let req = notify_request("POST", Some("Bearer  secret"), Bytes::new());
-        let resp = handle(client, req).await.unwrap();
-        assert_eq!(resp.status(), hyper::StatusCode::OK);
     }
 }
