+++ resource/ckb-miner.toml
@@ -61,6 +61,7 @@ poll_interval = 1000
 
 # Optional bearer token to authenticate block-template notifications.
 # Must match block_assembler.notify_auth_token in ckb.toml.
+# Must be non-empty without leading/trailing whitespace, otherwise ckb-miner refuses to start.
 # auth_token = "change-me-to-a-long-random-secret"
 
 [[miner.workers]]
