+++ resource/ckb.toml
@@ -273,6 +273,7 @@ block_uncles_cache_size    = 30
 # notify_scripts = ["your_notify_scripts.sh"]
 # # Optional bearer token to authenticate notify requests to ckb-miner.
 # # Must match miner.client.auth_token in ckb-miner.toml.
+# # Must be non-empty without leading/trailing whitespace, otherwise the node refuses to start.
 # notify_auth_token = "change-me-to-a-long-random-secret"
 #
 # # CKB built-in indexer/rich-indexer settings.
